add authentication with django-social-auth

+++ myyum/rpm/views.py
@@ -1,4 +1,5 @@
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db import IntegrityError
 from django.shortcuts import get_object_or_404, redirect, render_to_response
@@ -7,16 +8,25 @@ from django.template import RequestContext
 from myyum.rpm.models import *
 from myyum.rpm.forms import *
 
+@login_required
 def repository_index(request):
-    repos = Repository.objects.all()
+    repos = Repository.objects.filter(owner=request.user)
     return render_to_response("repo_index.html", dict(repos=repos), context_instance=RequestContext(request))
 
 
+@login_required
 def repository_view(request, repository_id):
     repo = get_object_or_404(Repository, id=repository_id)
+    
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
+    
     return render_to_response("repo_view.html", dict(repo=repo), context_instance=RequestContext(request))
 
-    
+
+@login_required    
 def repository_create(request):
     form = RepositoryForm(request.POST or None)
     
@@ -34,9 +44,15 @@ def repository_create(request):
         return render_to_response("repo_create.html", dict(form=form), context_instance=RequestContext(request))
 
 
+@login_required
 def repository_edit(request, repository_id):
     repo = get_object_or_404(Repository, id=repository_id)
     form = RepositoryForm(request.POST or None, instance=repo)
+    
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
     
     if form.is_valid():
         form.save()
@@ -46,8 +62,15 @@ def repository_edit(request, repository_id):
     else:
         return render_to_response("repo_edit.html", dict(form=form), context_instance=RequestContext(request))
 
+
+@login_required
 def repository_delete(request, repository_id):
     repo = get_object_or_404(Repository, id=repository_id)
+    
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
     
     if request.method == 'POST':
         repo.delete()
@@ -58,10 +81,15 @@ def repository_delete(request, repository_id):
         return render_to_response("repo_delete.html", dict(repo=repo), context_instance=RequestContext(request))
 
 
+@login_required
 def package_upload(request, repository_id):
     repo = get_object_or_404(Repository, id=repository_id)
-    
     form = PackageUploadForm(repo, request.POST or None, request.FILES or None)
+    
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
     
     if form.is_valid():
         try:
@@ -76,16 +104,28 @@ def package_upload(request, repository_id):
         return render_to_response("pkg_upload.html", dict(repo=repo, form=form), context_instance=RequestContext(request))
 
 
+@login_required
 def package_view(request, repository_id, package_id):
     repo = get_object_or_404(Repository, id=repository_id)
     package = get_object_or_404(RPMPackage, id=package_id)
     
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
+    
     return render_to_response("pkg_view.html", dict(repo=repo, pkg=package), context_instance=RequestContext(request))
 
 
+@login_required
 def package_delete(request, repository_id, package_id):
     repo = get_object_or_404(Repository, id=repository_id)
     pkg = get_object_or_404(RPMPackage, id=package_id)
+
+    # check rights
+    if not repo.owner == request.user:
+        messages.add_message(request, messages.WARNING, "Tried to access other users repository")
+        return redirect('myyum.rpm.views.repository_index')
 
     if request.method == 'POST':
         pkg.delete()
